add 1.2 task and fix tests
