+++ src/test/resources/floorData.csv
@@ -1,4 +1,4 @@
 brackets,expectedResult
-'((()',2
-'((()))))',-2
-'(())()',0
+"((()",2
+"((()))))",-2
+"(())()",0
